refactor: split render.go into per-platform packages
- cloudrun/ — Cloud Run renderer, owns secrets annotation
- k8s/ — Kubernetes renderer, owns hash suffixes and security context
- terraform/ — Terraform renderer, fully standalone
- internal/ — shared structural helpers (container, volumes, yaml marshal)

Each platform package is an independent go_library with its own tests.
