+++ devtools/bifrost/terraform/testdata/service.yaml
@@ -1,0 +1,33 @@
+apiVersion: bifrost.apotema.cloud/v1alpha1
+kind: Service
+metadata:
+  name: registry
+spec:
+  image: registry
+  args:
+    - --upstream=ghcr.io
+    - --repository-prefix=arkeros/senku
+  port: 8080
+  resources:
+    requests:
+      cpu: 250m
+      memory: 256Mi
+    limits:
+      cpu: 1000m
+      memory: 256Mi
+  secretFiles:
+    - secret: registry-env
+      path: /run/secrets/env.json
+  probes:
+    startupPath: /v2/
+    livenessPath: /v2/
+  autoscaling:
+    min: 0
+    max: 3
+  gcp:
+    projectId: senku-prod
+    projectNumber: "874944788122"
+    region: europe-west3
+    cloudRun:
+      ingress: all
+  kubernetes: {}
